celery task error fixes
response dictionary from bulk fetch was saved as weather_dict but for mdb saves it was never exploded to take resp_dict for particular weather type. This issue is now fixed and tested for bulk datis, metar and taf fetch and mdb save. Celery test through docker service is the next step to verify tasks are failing.